Add files via upload
Some countries in the dengue dataset had many NaN cells, so instead of conforming to standard preprocessing techniques and removing all of these cells, seasonal decomposition methods were used to fill in the missing data. Triple Exponential Smoothing, also known as the Holt-Winters method, was used to account for the trend, seasonality, and residual components of the dengue data to predict the missing values.